+++ parsers/LR0/types.h
@@ -1,5 +1,7 @@
 typedef pair< char,string>  ProdType;
 typedef vector< ProdType > Grammar;
 typedef vector< ProdType > StateType;
-
 typedef vector<string> RawGrammar ;
+
+typedef vector<string> ParseRowType;
+typedef vector<ParseRowType> ParseTableType;
